+++ scanpy/neighbors/_enums.py
